fixed damaging projetiles by adding things to the 'Battle' layer. Still in progress is the aura ability and associated pattern for AbilityEffects
