+++ Assets/Scenes/UIScene.unity
@@ -38,7 +38,7 @@ RenderSettings:
   m_ReflectionIntensity: 1
   m_CustomReflection: {fileID: 0}
   m_Sun: {fileID: 705507994}
-  m_IndirectSpecularColor: {r: 0.44657898, g: 0.4964133, b: 0.5748178, a: 1}
+  m_IndirectSpecularColor: {r: 0.44657844, g: 0.49641222, b: 0.57481694, a: 1}
   m_UseRadianceAmbientProbe: 0
 --- !u!157 &3
 LightmapSettings:
@@ -134,7 +134,8 @@ GameObject:
   - component: {fileID: 247324052}
   - component: {fileID: 247324051}
   - component: {fileID: 247324050}
-  m_Layer: 0
+  - component: {fileID: 247324056}
+  m_Layer: 9
   m_Name: FakePlayer
   m_TagString: Untagged
   m_Icon: {fileID: 0}
@@ -240,6 +241,18 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 4
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &247324056
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 247324049}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: dd9c4570326667445a168e0212cf4883, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1 &587326200
 GameObject:
   m_ObjectHideFlags: 0
@@ -510,7 +523,7 @@ GameObject:
   - component: {fileID: 762576036}
   - component: {fileID: 762576040}
   - component: {fileID: 762576041}
-  m_Layer: 0
+  m_Layer: 9
   m_Name: CuteP
   m_TagString: Untagged
   m_Icon: {fileID: 0}
@@ -696,6 +709,12 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 7
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!4 &1222314705 stripped
+Transform:
+  m_CorrespondingSourceObject: {fileID: 2045603164783674830, guid: bab3af7e4f87e304fb5424ca02ad780a,
+    type: 3}
+  m_PrefabInstance: {fileID: 2045603164955066954}
+  m_PrefabAsset: {fileID: 0}
 --- !u!1 &1385117675
 GameObject:
   m_ObjectHideFlags: 0
@@ -769,6 +788,51 @@ CanvasRenderer:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1385117675}
   m_CullTransparentMesh: 0
+--- !u!1 &1526772182 stripped
+GameObject:
+  m_CorrespondingSourceObject: {fileID: 2045603164883527681, guid: bab3af7e4f87e304fb5424ca02ad780a,
+    type: 3}
+  m_PrefabInstance: {fileID: 2045603164955066954}
+  m_PrefabAsset: {fileID: 0}
+--- !u!114 &1526772189
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1526772182}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: dd9c4570326667445a168e0212cf4883, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+--- !u!114 &1526772190
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1526772182}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: e746ad0d996647d40a8c690d73eae1e5, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+--- !u!114 &1526772191
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1526772182}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: f79b8a08dbef1b64cbac6dd8cfc8f3bf, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  projectile: {fileID: 0}
+  projectileSpawnPoint: {fileID: 1222314705}
+  projectileForce: 300
 --- !u!1 &1565753367
 GameObject:
   m_ObjectHideFlags: 0
@@ -885,6 +949,17 @@ PrefabInstance:
       propertyPath: parentPanel
       value: 
       objectReference: {fileID: 1385117675}
+    - target: {fileID: 2045603164778340387, guid: bab3af7e4f87e304fb5424ca02ad780a,
+        type: 3}
+      propertyPath: abilities.Array.data[1]
+      value: 
+      objectReference: {fileID: 11400000, guid: a4c004a701c743449adf67a75f9a3947,
+        type: 2}
+    - target: {fileID: 2045603164783674825, guid: bab3af7e4f87e304fb5424ca02ad780a,
+        type: 3}
+      propertyPath: m_Layer
+      value: 9
+      objectReference: {fileID: 0}
     - target: {fileID: 2045603164883498019, guid: bab3af7e4f87e304fb5424ca02ad780a,
         type: 3}
       propertyPath: m_LocalPosition.x
@@ -939,6 +1014,11 @@ PrefabInstance:
         type: 3}
       propertyPath: m_LocalEulerAnglesHint.z
       value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2045603164883527681, guid: bab3af7e4f87e304fb5424ca02ad780a,
+        type: 3}
+      propertyPath: m_Layer
+      value: 9
       objectReference: {fileID: 0}
     - target: {fileID: 2045603164883527683, guid: bab3af7e4f87e304fb5424ca02ad780a,
         type: 3}
