+++ Assets/Scripts/Abilities/Implemented Abilities/Aura/FireAura-DamageOverTime.asset
@@ -1,0 +1,16 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 7590c0167d958fb449f5205bb83cf5db, type: 3}
+  m_Name: FireAura-DamageOverTime
+  m_EditorClassIdentifier: 
+  TypeOfEffect: 0
+  DamagePerSecond: 5
