+++ Assets/Scripts/Abilities/Implemented Abilities/Aura/FireAura-DamageOverTime.asset.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 5c6255eeef44c874db17daa93af49bd9
-folderAsset: yes
-DefaultImporter:
+guid: 27c36101f2181af47aa033d16c65b3d1
+NativeFormatImporter:
   externalObjects: {}
+  mainObjectFileID: 11400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
