+++ Assets/Scripts/Abilities/Implemented Abilities/Aura/FireAura.asset
@@ -10,21 +10,21 @@ MonoBehaviour:
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: 2a69b08340238cf46ab5858f82f3dada, type: 3}
-  m_Name: Force Pulse
+  m_Name: FireAura
   m_EditorClassIdentifier: 
-  abilityName: Force Pulse
-  baseCooldown: 20
+  abilityName: Fire Aura
+  baseCooldown: 10
   castTime: 0
   targetKnockbackForce: 0
   casterKnockbackForce: 0
   maxStacks: 1
   damage: 0
-  abilitySprite: {fileID: 0}
+  abilitySprite: {fileID: 21300000, guid: 9866a92691696b346901281f2b329034, type: 3}
   effect: {fileID: 0}
   abilitySound: {fileID: 0}
   auraDuration: 10
-  auraRadius: 2
+  auraRadius: 5
   auraEffects:
-  - {fileID: 11400000, guid: c4390ce8e0e48f94196c14f865685892, type: 2}
-  doesPulse: 1
-  pulseDelay: 2
+  - {fileID: 11400000, guid: 27c36101f2181af47aa033d16c65b3d1, type: 2}
+  doesPulse: 0
+  pulseDelay: 0
