+++ Assets/Scripts/Abilities/Implemented Abilities/Aura/FireAura.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a4c004a701c743449adf67a75f9a3947
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
