+++ Assets/Scripts/Abilities/ProjectileAbilityCollisionHandler.cs
@@ -15,7 +15,11 @@ public class ProjectileAbilityCollisionHandler : MonoBehaviour
 
     private void OnCollisionEnter(Collision other)
     {
-        this.onCollisionFunction(other, this.gameObject);
-        Destroy(this, .25f);
+        if (other.transform.gameObject.layer == LayerMask.NameToLayer("Battle"))
+        {
+            Debug.Log("OnCollisionEnter inside ProjectileAbilityCollisionHandler");
+            this.onCollisionFunction(other, this.gameObject);
+            Destroy(this.gameObject, .25f);
+        }
     }
 }
