+++ Assets/Scripts/Health.cs
@@ -1,8 +1,9 @@
-﻿using System.Collections;
+﻿using Assets.Scripts;
+using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-public class Health : MonoBehaviour
+public class Health : MonoBehaviour, IHealth
 {
     public float maxHealth = 100;
     public float currentHealth;
@@ -16,6 +17,7 @@ public class Health : MonoBehaviour
 
     public void TakeDamage(float amount)
     {
+        Debug.Log("In take damage for Normal Health");
         ChangeHealth(-amount);
     }
 
@@ -26,6 +28,7 @@ public class Health : MonoBehaviour
 
     public virtual void ChangeHealth(float delta)
     {
+        Debug.Log("In ChangeHealth for Normal Health");
         this.currentHealth += delta;
         DestroyOnDeath();
     }
