+++ Assets/Scripts/IHealth.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 9cbb4c7285430c74c94358257ddcd582
+guid: f1543046470c7ac4589a013f328256f1
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
