+++ Assets/Scripts/PlayerScripts/AuraAbilityHandler.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e746ad0d996647d40a8c690d73eae1e5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
