+++ Assets/Scripts/PlayerScripts/PlayerManager.cs
@@ -7,10 +7,12 @@ public class PlayerManager : MonoBehaviour
     //Delegate definitions
     public delegate void HitByAbilityDelegate(AbilityBase ability);
     public delegate void PlayerDeathDelegate();
+    public delegate void PlayerTookDamage();
 
     //Delegate references
     public HitByAbilityDelegate onPlayerHit;
     public PlayerDeathDelegate onPlayerDeath;
+    public PlayerTookDamage onPlayerDamaged;
 
     //Handlers
     public void HandlePlayerHitByAbility(AbilityBase someAbility)
@@ -21,5 +23,9 @@ public class PlayerManager : MonoBehaviour
     public void HandlePlayerDeath()
     {
         onPlayerDeath?.Invoke();
+    }
+    public void HandlePlayerTookDamage()
+    {
+        onPlayerDamaged?.Invoke();
     }
 }
