+++ Assets/Scripts/PlayerScripts/ProjectileAbilityHandler.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 6d21a831f77d86b4cb91d5026e64d08a
+guid: f79b8a08dbef1b64cbac6dd8cfc8f3bf
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
